rename ProductFeatureContext->CommonProductFeatureContext
